+++ src/aws/cis-1.4.0/tests/aws-cis-1.4.0-2.x.test.ts
@@ -3,9 +3,33 @@ import cuid from 'cuid'
 import { initRuleEngine } from '../../../utils/test'
 
 import Aws_CIS_140_211 from '../rules/aws-cis-1.4.0-2.1.1'
+import Aws_CIS_140_212 from '../rules/aws-cis-1.4.0-2.1.2'
 import Aws_CIS_140_213 from '../rules/aws-cis-1.4.0-2.1.3'
-import Aws_CIS_140_215 from '../rules/aws-cis-1.4.0-2.1.5'
+import Aws_CIS_140_215_1 from '../rules/aws-cis-1.4.0-2.1.5.1'
+import Aws_CIS_140_215_2 from '../rules/aws-cis-1.4.0-2.1.5.2'
+import Aws_CIS_140_221 from '../rules/aws-cis-1.4.0-2.2.1'
 import Aws_CIS_140_231 from '../rules/aws-cis-1.4.0-2.3.1'
+
+export interface Condition {
+  key: string
+  value: string[]
+}
+
+export interface Principal {
+  key: string
+  value: string[]
+}
+
+export interface Statement {
+  effect: string
+  action: string[]
+  principal: Principal[]
+  condition: Condition[]
+}
+
+export interface Policy {
+  statement: Statement[]
+}
 
 export interface QueryawsRdsDbInstance {
   id: string
@@ -17,17 +41,27 @@ export interface EncryptionRule {
 }
 export interface QueryawsS3 {
   id: string
+  policy?: Policy
   versioning?: string
   mfa?: string
   blockPublicAcls?: string
   ignorePublicAcls?: string
   blockPublicPolicy?: string
   restrictPublicBuckets?: string
+  accountLevelBlockPublicAcls?: string,
+  accountLevelIgnorePublicAcls?: string,
+  accountLevelBlockPublicPolicy?: string,
+  accountLevelRestrictPublicBuckets?: string
   encrypted?: string
   encryptionRules?: EncryptionRule[]
 }
+export interface QueryawsEbs {
+  id: string
+  encrypted: boolean
+}
 export interface CIS2xQueryResponse {
   queryawsS3?: QueryawsS3[]
+  queryawsEbs?: QueryawsEbs[]
   queryawsRdsDbInstance?: QueryawsRdsDbInstance[]
 }
 
@@ -88,6 +122,102 @@ describe('CIS Amazon Web Services Foundations: 1.4.0', () => {
     })
   })
 
+  describe('AWS CIS 2.1.2 Ensure S3 Bucket Policy allows HTTPS requests', () => {
+    const getTestRuleFixture = (
+      effect: string,
+      action: string,
+      principal: Principal,
+      condition: Condition
+    ): CIS2xQueryResponse => {
+      return {
+        queryawsS3: [
+          {
+            id: cuid(),
+            policy: {
+              statement: [
+                {
+                  effect,
+                  action: [action],
+                  principal: [principal],
+                  condition: [condition],
+                },
+              ],
+            },
+          },
+        ],
+      }
+    }
+
+    // Act
+    const testRule = async (
+      data: CIS2xQueryResponse,
+      expectedResult: Result
+    ): Promise<void> => {
+      // Act
+      const [processedRule] = await rulesEngine.processRule(
+        Aws_CIS_140_212 as Rule,
+        { ...data }
+      )
+
+      // Asserts
+      expect(processedRule.result).toBe(expectedResult)
+    }
+
+    test('No Security Issue when S3 bucket policies only allow requests that use HTTPS', async () => {
+      const principal: Principal = {
+        key: 'AWS',
+        value: ['*'],
+      }
+      const condition: Condition = {
+        key: 'aws:SecureTransport',
+        value: ['false'],
+      }
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'Deny',
+        's3:*',
+        principal,
+        condition
+      )
+
+      await testRule(data, Result.PASS)
+    })
+
+    test('Security Issue when S3 bucket policy does not have SecureTransport enabled', async () => {
+      const principal: Principal = {
+        key: 'AWS',
+        value: ['arn:aws:iam::111122223333:root'],
+      }
+      const condition: Condition = {
+        key: 'aws:SecureTransport',
+        value: ['false'],
+      }
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'Allow',
+        's3:*',
+        principal,
+        condition
+      )
+
+      await testRule(data, Result.FAIL)
+    })
+
+    test('Security Issue when S3 bucket policy have SecureTransport enabled but grants permission to any public anonymous users', async () => {
+      const principal: Principal = { key: '', value: ['*'] }
+      const condition: Condition = {
+        key: 'aws:SecureTransport',
+        value: ['true'],
+      }
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'Allow',
+        's3:*',
+        principal,
+        condition
+      )
+
+      await testRule(data, Result.FAIL)
+    })
+  })
+
   describe('AWS CIS 2.1.3 Ensure MFA Delete is enable on S3 buckets', () => {
     const getTestRuleFixture = (
       versioning: string,
@@ -143,7 +273,103 @@ describe('CIS Amazon Web Services Foundations: 1.4.0', () => {
     })
   })
 
-  describe('AWS CIS 2.1.5 Ensure that S3 Buckets are configured with Block public access (bucket settings)', () => {
+  describe('AWS CIS 2.1.5.1 Ensure that S3 Buckets are configured with Block public access (account settings)', () => {
+    const getTestRuleFixture = (
+      accountLevelBlockPublicAcls: string,
+      accountLevelIgnorePublicAcls: string,
+      accountLevelBlockPublicPolicy: string,
+      accountLevelRestrictPublicBuckets: string,
+    ): CIS2xQueryResponse => {
+      return {
+        queryawsS3: [
+          {
+            id: cuid(),
+            accountLevelBlockPublicAcls,
+            accountLevelIgnorePublicAcls,
+            accountLevelBlockPublicPolicy,
+            accountLevelRestrictPublicBuckets,
+          },
+        ],
+      }
+    }
+
+    // Act
+    const testRule = async (
+      data: CIS2xQueryResponse,
+      expectedResult: Result
+    ): Promise<void> => {
+      // Act
+      const [processedRule] = await rulesEngine.processRule(
+        Aws_CIS_140_215_1 as Rule,
+        { ...data }
+      )
+
+      // Asserts
+      expect(processedRule.result).toBe(expectedResult)
+    }
+
+    test('No Security Issue when S3 Account Level is configured with Block public access', async () => {
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'Yes',
+        'Yes',
+        'Yes',
+        'Yes'
+      )
+      await testRule(data, Result.PASS)
+    })
+
+    test('Security Issue when S3 Account Level is not configured with Block public access', async () => {
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'No',
+        'No',
+        'No',
+        'No'
+      )
+      await testRule(data, Result.FAIL)
+    })
+
+    test('Security Issue when S3 Account Level have a Block public access with blockPublicAcls set to No', async () => {
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'No',
+        'Yes',
+        'Yes',
+        'Yes'
+      )
+      await testRule(data, Result.FAIL)
+    })
+
+    test('Security Issue when S3 Account Level have a Block public access with ignorePublicAcls set to No', async () => {
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'Yes',
+        'No',
+        'Yes',
+        'Yes'
+      )
+      await testRule(data, Result.FAIL)
+    })
+
+    test('Security Issue when S3 Account Level have a Block public access with blockPublicPolicy set to No', async () => {
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'Yes',
+        'Yes',
+        'No',
+        'Yes'
+      )
+      await testRule(data, Result.FAIL)
+    })
+
+    test('Security Issue when S3 Account Level have a Block public access with restrictPublicBuckets set to No', async () => {
+      const data: CIS2xQueryResponse = getTestRuleFixture(
+        'Yes',
+        'Yes',
+        'Yes',
+        'No'
+      )
+      await testRule(data, Result.FAIL)
+    })
+  })
+
+  describe('AWS CIS 2.1.5.2 Ensure that S3 Buckets are configured with Block public access (bucket settings)', () => {
     const getTestRuleFixture = (
       blockPublicAcls: string,
       ignorePublicAcls: string,
@@ -170,7 +396,7 @@ describe('CIS Amazon Web Services Foundations: 1.4.0', () => {
     ): Promise<void> => {
       // Act
       const [processedRule] = await rulesEngine.processRule(
-        Aws_CIS_140_215 as Rule,
+        Aws_CIS_140_215_2 as Rule,
         { ...data }
       )
 
@@ -235,6 +461,44 @@ describe('CIS Amazon Web Services Foundations: 1.4.0', () => {
         'Yes',
         'No'
       )
+      await testRule(data, Result.FAIL)
+    })
+  })
+
+  describe('AWS CIS 2.2.1 Ensure EBS volume encryption is enabled', () => {
+    const getTestRuleFixture = (encrypted: boolean): CIS2xQueryResponse => {
+      return {
+        queryawsEbs: [
+          {
+            id: cuid(),
+            encrypted,
+          },
+        ],
+      }
+    }
+
+    // Act
+    const testRule = async (
+      data: CIS2xQueryResponse,
+      expectedResult: Result
+    ): Promise<void> => {
+      // Act
+      const [processedRule] = await rulesEngine.processRule(
+        Aws_CIS_140_221 as Rule,
+        { ...data }
+      )
+
+      // Asserts
+      expect(processedRule.result).toBe(expectedResult)
+    }
+
+    test('No Security Issue when EBS volume encryption is enabled', async () => {
+      const data: CIS2xQueryResponse = getTestRuleFixture(true)
+      await testRule(data, Result.PASS)
+    })
+
+    test('Security Issue when EBS volume encryption is not enabled', async () => {
+      const data: CIS2xQueryResponse = getTestRuleFixture(false)
       await testRule(data, Result.FAIL)
     })
   })