+++ src/aws/cis-1.5.0/CHANGELOG.md
@@ -1,5 +1,20 @@
-### Bug Fixes
+# @cloudgraph/policy-pack-aws-cis-1.5.0-v1.0.0-alpha.1 (2022-12-14)
 
 
 ### Features
 
+* **CG-1280:** add aws cis 1.5 iam rules ([9bac0ec](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/9bac0ec8ceb27ca2f77a85f4d11154f311a75b12))
+* **CG-1281:** add aws cis 1.5 logging rules ([1bf5c08](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/1bf5c08273a5ea41f842ce9d8be48802a50a1bcf))
+* **CG-1282:** add support monitoring rules ([46c9483](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/46c94833fae665b9ab12482a247ff5a63d03fc42))
+* **CG-1284:** add networking rule support ([b758b72](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/b758b72735948f823b131b30b82be3794b1d2215))
+* **CG-1285:** add aws cis 150 5.3 ([6528999](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/65289997ec7a51f07b3a4392db7d487206418828))
+* **CG-1286:** add storage rules support ([0934441](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/09344413a18f1f6289e1cd02f8d594983b5dde46))
+* **CG-1287:** add aws cis 2.3.2 support ([e6ee33f](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/e6ee33fec483681e6539df01a3aa89359073e11b))
+* **CG-1288:** add aws cis 233 support ([81359bf](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/81359bf30d66f3d849f1c6b8b3b6cd9ab1356842))
+* **CG-1289:** add aws cis 1.5.0 2.4.1 ([998bf70](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/998bf7013e8c3acbc8633ced5e449464a6ee3d26))
+* **CG-1290:** create boilerplate for aws cis 1.5.0 ([9c1d6f3](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/9c1d6f392a03b506424e74682caca85ff9188238))
+
+### Bug Fixes
+
+
+### Features