+++ src/azure/pci-dss-3.2.1/CHANGELOG.md
@@ -1,3 +1,28 @@
+# [@cloudgraph/policy-pack-azure-pci-dss-3.2.1-v1.2.0-alpha.1](https://github.com/cloudgraphdev/cloudgraph-policy-packs/compare/@cloudgraph/policy-pack-azure-pci-dss-3.2.1@1.1.0...@cloudgraph/policy-pack-azure-pci-dss-3.2.1@1.2.0-alpha.1) (2022-12-14)
+
+
+### Bug Fixes
+
+* **pnpm:** using semantic-release-pnpm ([eb9f3f2](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/eb9f3f22e85375b79be205c62adc09aa60628343))
+
+
+### Features
+
+* **CG-1164:** add Activity Log Retention ([ec03e27](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/ec03e276159daa8a57da018071e32b73d6aa91ab))
+* **CG-1164:** add azure pci sql server auditing enabled check ([a9e34a7](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/a9e34a717d8f864fa8c98a33ccb36a3420dd8098))
+* **CG-1165:** add azure monitoring rule ([4383c3c](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/4383c3ca29a5a85a52a206590f71545ee09d713b))
+* **CG-1165:** add azure pci monitor log profile rule ([a72f81c](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/a72f81c49c4a30704d5ad641de290d52114282d8))
+* **CG-1165:** rule name update ([4ca8eb7](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/4ca8eb7d53f6e99a9655ff7bb86eb889ab027115))
+* **CG-1165:** update README ([11e2604](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/11e26043b301c0dfc7235552870d4f88b472a6a2))
+* **CG-1168:** add azure networking 3 rule ([6d3925d](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/6d3925dfbd6c06e717a3ce650b594a8e6baa0df0))
+* **CG-1169:** add azure user-check-1 rule ([2f1ef53](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/2f1ef537624f30b644d8b4ca15881a052f3ee007))
+* **CG-1173:** add azure WAF enabled check ([a6a8b31](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/a6a8b31b40d4711e043ac2e717df2ee5491f1c5a))
+* **CG-1174:** fix monitoring check 5 ([4b3c357](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/4b3c357096cc50d729b542b4204fb7c8a2bcf88c))
+* **CG-1174:** support azure pci monitoring rules ([3b86f5c](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/3b86f5cd017409e11b5e4d5d1aa6867349f81a9b))
+* **CG-1175:** add azure encryption transit rule check ([e6f01d4](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/e6f01d4bcf02d5d7f243ad96fec4302e75a58e9b))
+* **CG-1176:** add network access rules and policies version check rules ([d4e1604](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/d4e16043affc4553d4fb56d05f908461ff439123))
+* **CG-1176:** README update ([bfde2ed](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/bfde2edced9fab3210c9a69a740a1659b9208bd1))
+
 # [@cloudgraph/policy-pack-azure-pci-dss-3.2.1-v1.1.0](https://github.com/cloudgraphdev/cloudgraph-policy-packs/compare/@cloudgraph/policy-pack-azure-pci-dss-3.2.1@1.0.3...@cloudgraph/policy-pack-azure-pci-dss-3.2.1@1.1.0) (2022-08-01)